게시글 삭제 성공 - AJAX 요청에 대한 공통 Dto, Exception File 생성

+++ src/main/java/com/cos/prjchr/web/BoardController.java
@@ -22,9 +22,11 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.cos.prjchr.domain.board.Board;
 import com.cos.prjchr.domain.board.BoardRepository;
 import com.cos.prjchr.domain.user.User;
+import com.cos.prjchr.handler.ex.MyAsyncNotFoundException;
 import com.cos.prjchr.handler.ex.MyNotFoundException;
 import com.cos.prjchr.util.Script;
 import com.cos.prjchr.web.dto.BoardSaveReqDto;
+import com.cos.prjchr.web.dto.CMRespDto;
 
 import lombok.RequiredArgsConstructor;
 
@@ -38,9 +40,33 @@ public class BoardController {
 
 	// DELETE FROM board WHERE id = ?
 	@DeleteMapping("/board/{id}")
-	public @ResponseBody String deleteById(@PathVariable int id) {
-		boardRepository.deleteById(id);
-		return "ok"; // @ResponseBody 데이터 리턴!! String = text/plain
+//	public @ResponseBody String deleteById(@PathVariable int id) {
+//		boardRepository.deleteById(id);
+//		return "ok"; // @ResponseBody 데이터 리턴!! String = text/plain
+//	}
+	public @ResponseBody CMRespDto<String> deleteById(@PathVariable int id) {
+
+		// 인증이 된 사람만 함수 접근 가능!! (로그인 된 사람)
+		User principal = (User) session.getAttribute("principal");
+		if(principal == null) {
+			throw new MyAsyncNotFoundException("인증이 되지 않았습니다.");
+		}
+
+		// 권한이 있는 사람만 함수 접근 가능(principal.id == {id})
+		Board boardEntity = boardRepository.findById(id)
+			.orElseThrow(()-> new MyAsyncNotFoundException("해당글을 찾을 수 없습니다."));
+		if(principal.getId() != boardEntity.getUser().getId()) {
+			throw new MyAsyncNotFoundException("해당글을 삭제할 권한이 없습니다.");
+		}
+
+		try {
+			boardRepository.deleteById(id); // 오류 발생??? (id가 없으면) 
+		} catch (Exception e) {
+			throw new MyAsyncNotFoundException(id+"를 찾을 수 없어서 삭제할 수 없어요.");
+		}
+
+
+		return new CMRespDto<String>(1, "성공", null); // @ResponseBody 데이터 리턴!! String
 	}
 	
 //
